+++ packages/component-library/src/components/templates/ApplicationLayout/ApplicationLayout.css.ts
@@ -1,6 +1,6 @@
 import { globalStyle, style } from '@vanilla-extract/css';
 import { rem } from 'polished';
-import { globalHeadingStyle, responsiveText } from 'src/style/typography/typography.css';
+import { globalHeadingStyle, responsiveText } from '../../../style/typography/typography.css';
 import { responsiveStyle } from '../../../style';
 import { breakpoints } from '../../../style/breakpoints';
 import { vars } from '../../../style/themes/vars.css';
@@ -81,10 +81,10 @@ globalStyle('header,footer', {
   ...bodyStyle,
 });
 
-// globalStyle('header', {
-//   background: vars.backgroundColor.heading,
-//   color: vars.foregroundColor.heading,
-// });
+globalStyle('header', {
+  background: vars.backgroundColor.header,
+  color: vars.foregroundColor.header,
+});
 
 globalStyle('footer', {
   background: vars.backgroundColor.footer,