+++ packages/website/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @cutting/website
 
+## 4.27.13
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @cutting/component-library@5.3.8
+
+## 4.27.12
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @cutting/component-library@5.3.7
+
+## 4.27.11
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @cutting/component-library@5.3.6
+
 ## 4.27.10
 
 ### Patch Changes